pull the outstation client and server pieces apart
