+++ dnp3/src/tcp/outstation/mod.rs
@@ -1,0 +1,5 @@
+mod client;
+mod server;
+
+pub use client::*;
+pub use server::*;
